Can't read values from external sources
Closes #97

vault-cli has been designed read and write all the values in a "value" key.
With the kv1 secret engine, vault set a b will put this kind of secret in the vault :

```
data:
  value: b
```

But some values from external sources (e.g. the rabbitmq secret engine) have this kind of secret :
```
data:
  username: foo
  password: bar
```

Here, if the secret does not contain a single "value" key, we return
directly the whole kv dict.